# backend/src/routes/auth.js
const express = require('express');
const bcrypt = require('bcryptjs');
const crypto = require('crypto');
const { v4: uuid } = require('uuid');
const { q, one } = require('../config/db');
const { sign, signRefresh, verifyRefresh } = require('../middleware/auth');
const { v, schemas } = require('../middleware/validate');
const { audit } = require('../middleware/audit');
const r = express.Router();

// Issue an access + refresh pair; refresh stored hashed server-side (rotation).
async function issuePair(u) {
  const token = sign(u);
  const refreshToken = signRefresh(u);
  const th = crypto.createHash('sha256').update(refreshToken).digest('hex');
  await q(`INSERT INTO refresh_tokens (id, user_id, token_hash, expires_at)
           VALUES (:id, :who, :th, SYSTIMESTAMP + INTERVAL '30' DAY)`,
    { id: uuid(), who: u.id, th });
  return { token, refreshToken };
}

// POST /api/auth/register — student/alumni self-onboarding (coordinator seeded via SQL*Plus)
r.post('/register', v(schemas.register), async (req, res) => {
  const { email, password, full_name, role, languages } = req.body;
  try {
    const exists = await one('SELECT id FROM users WHERE email = :email', { email: email.toLowerCase() });
    if (exists) return res.status(409).json({ error: 'Email already registered' });
    const hash = await bcrypt.hash(password, 12); // bcrypt-12, never MD5/SHA1
    const id = uuid();
    await q(`INSERT INTO users (id, email, password_hash, role, full_name, languages, consent_given)
             VALUES (:id, :email, :hash, :role, :name, :langs, 1)`,
      { id, email: email.toLowerCase(), hash, role, name: full_name, langs: JSON.stringify(languages) });
    if (role === 'student') {
      await q(`INSERT INTO student_profiles (user_id, enrollment_no, department, study_year, goals, domain_interests)
               VALUES (:id, :enr, 'IT', 3, :goals, :domains)`,
        { id, enr: 'ENR-' + id.slice(0, 8).toUpperCase(), goals: JSON.stringify({ primary: 'placement' }), domains: JSON.stringify(['Web Dev']) });
    } else {
      await q(`INSERT INTO alumni_profiles (user_id, graduation_year, expertise_tags, max_mentees)
               VALUES (:id, 2022, :tags, 5)`, { id, tags: JSON.stringify(['Web Dev', 'DevOps']) });
    }
    await audit(id, 'AUTH.REGISTER', 'users', id, { role }, req.ip);
    res.status(201).json({ id, ...(await issuePair({ id, role, email })) });
  } catch (e) {
    if (e.errorNum === 1) return res.status(409).json({ error: 'Email already registered' });
    res.status(500).json({ error: 'Register failed' });
  }
});

// POST /api/auth/login — generic errors (no user enumeration) + DB-backed lockout
r.post('/login', v(schemas.login), async (req, res) => {
  const { email, password } = req.body;
  const mail = String(email).toLowerCase();
  try {
    const lock = await one('SELECT mentor_sec.fn_is_locked(:email) AS locked FROM DUAL', { email: mail });
    if (lock && Number(lock.locked) === 1) {
      await audit(null, 'AUTH.LOCKED', 'users', null, { email: mail }, req.ip);
      return res.status(429).json({ error: 'Too many failed attempts. Try again in 15 minutes.' });
    }
    const u = await one('SELECT id, email, password_hash, role, full_name FROM users WHERE email = :email AND is_active = 1', { email: mail });
    const ok = u && await bcrypt.compare(password, u.password_hash);
    if (!ok) {
      // constant-shape failure: attacker learns nothing (no "user not found" vs "wrong password")
      await q('BEGIN mentor_sec.proc_reg_fail(:email, :ip); END;', { email: mail, ip: req.ip || null });
      await audit(null, 'AUTH.FAIL', 'users', null, { email: mail }, req.ip);
      return res.status(401).json({ error: 'Invalid credentials' });
    }
    await q('BEGIN mentor_sec.proc_reg_success(:email); END;', { email: mail });
    await audit(u.id, 'AUTH.LOGIN', 'users', u.id, {}, req.ip);
    res.json({ ...(await issuePair(u)), user: { id: u.id, role: u.role, full_name: u.full_name, email: u.email } });
  } catch {
    res.status(500).json({ error: 'Login failed' });
  }
});

// GET /api/auth/me — never returns password_hash (column allowlist)
const { auth } = require('../middleware/auth');
r.get('/me', auth, async (req, res) => {
  const me = await one('SELECT id, email, role, full_name, languages FROM users WHERE id = :id', { id: req.user.id });
  res.json(me || null);
});
// POST /api/auth/forgot — always 200 (no enumeration). Demo-mode: token returned
// so the flow works without SMTP; in production, email it instead of returning it.
r.post('/forgot', async (req, res) => {
  const { error, value } = require('../middleware/validate').schemas.login
    .fork('password', (s) => s.optional()).validate(req.body, { stripUnknown: true });
  if (error || !value.email) return res.json({ ok: true });
  const u = await one('SELECT id FROM users WHERE email = :email AND is_active = 1', { email: String(value.email).toLowerCase() });
  if (!u) { await audit(null, 'AUTH.FORGOT_MISS', 'users', null, {}, req.ip); return res.json({ ok: true }); }
  const token = uuid() + uuid().replace(/-/g, '');
  const th = crypto.createHash('sha256').update(token).digest('hex');
  await q(`INSERT INTO password_resets (id, user_id, token_hash, expires_at)
           VALUES (:id, :who, :th, SYSTIMESTAMP + INTERVAL '1' HOUR)`,
    { id: uuid(), who: u.id, th });
  await audit(u.id, 'AUTH.FORGOT', 'users', u.id, {}, req.ip);
  res.json({ ok: true, demo_token: token });
});
// POST /api/auth/reset — consume a forgot token, set a new password
r.post('/reset', async (req, res) => {
  const { token, password } = req.body || {};
  if (typeof token !== 'string' || !token) return res.status(400).json({ error: 'token required' });
  const { error } = require('joi').string().min(10).max(128).required().validate(password);
  if (error) return res.status(400).json({ error: 'Password needs 10+ chars' });
  const th = crypto.createHash('sha256').update(token).digest('hex');
  const row = await one(`SELECT id, user_id, expires_at, used FROM password_resets
                         WHERE token_hash = :th AND used = 0 AND expires_at > SYSTIMESTAMP`, { th });
  if (!row) return res.status(400).json({ error: 'Invalid or expired token' });
  const hash = await bcrypt.hash(password, 12);
  await q('UPDATE users SET password_hash = :h WHERE id = :id', { h: hash, id: row.user_id });
  await q('UPDATE password_resets SET used = 1 WHERE id = :id', { id: row.id });
  await q('DELETE FROM login_attempts WHERE email = (SELECT email FROM users WHERE id = :id)', { id: row.user_id });
  await audit(row.user_id, 'AUTH.RESET', 'users', row.user_id, {}, req.ip);
  res.json({ ok: true });
});
r.post('/logout', auth, async (req, res) => {
  const { refreshToken } = req.body || {};
  if (typeof refreshToken === 'string' && refreshToken) {
    const th = crypto.createHash('sha256').update(refreshToken).digest('hex');
    await q('UPDATE refresh_tokens SET revoked = 1 WHERE token_hash = :th AND user_id = :who', { th, who: req.user.id });
  }
  await audit(req.user.id, 'AUTH.LOGOUT', 'users', req.user.id, {}, req.ip);
  res.json({ ok: true });
});
// POST /api/auth/refresh — rotate refresh token, get a fresh pair (SaaS sessions)
r.post('/refresh', async (req, res) => {
  const { refreshToken } = req.body || {};
  if (typeof refreshToken !== 'string' || !refreshToken) return res.status(401).json({ error: 'Missing refresh token' });
  let p;
  try { p = verifyRefresh(refreshToken); } catch { return res.status(401).json({ error: 'Invalid refresh token' }); }
  const th = crypto.createHash('sha256').update(refreshToken).digest('hex');
  const row = await one(`SELECT id, user_id FROM refresh_tokens
                         WHERE token_hash = :th AND revoked = 0 AND expires_at > SYSTIMESTAMP`, { th });
  if (!row || row.user_id !== p.id) return res.status(401).json({ error: 'Invalid refresh token' });
  const u = await one('SELECT id, email, role FROM users WHERE id = :id AND is_active = 1', { id: row.user_id });
  if (!u) return res.status(401).json({ error: 'Invalid refresh token' });
  await q('UPDATE refresh_tokens SET revoked = 1 WHERE id = :id', { id: row.id });
  await audit(u.id, 'AUTH.REFRESH', 'users', u.id, {}, req.ip);
  res.json(await issuePair(u));
});
module.exports = r;