# openapi.yaml
openapi: 3.0.3
info: { title: MentorSetu API (BIT-05), version: 2.0.0, description: 'Explainable capacity-aware mentor matching. Served at /openapi.yaml. Auth: Bearer JWT.' }
servers: [{ url: http://localhost:4000 }]
components:
  securitySchemes: { bearer: { type: http, scheme: bearer, bearerFormat: JWT } }
...
security: [{ bearer: [] }]
paths:
...
  /api/auth/register: { post: { security: [], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email, password, full_name, role], properties: { email: { type: string }, password: { type: string, minLength: 10 }, full_name: { type: string }, role: { enum: [student, alumni] } } } } } }, responses: { '201': { description: '{id, token}' }, '400': { description: validation }, '409': { description: duplicate } } } }
  /api/auth/login: { post: { security: [], requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { email: { type: string }, password: { type: string } } } } } }, responses: { '200': { description: '{token, user}' }, '401': { description: invalid }, '429': { description: locked } } } }
...
  /api/auth/reset: { post: { security: [], responses: { '200': { description: password changed }, '400': { description: bad token/password } } } }
  /api/auth/refresh: { post: { security: [], responses: { '200': { description: rotated pair }, '401': { description: invalid } } } }
...